refactor(run-once): split planning phase orchestration
